WikiMine
This project was for the purpose of giving the client a method to use in searching two different document servers with one user interface. Client attempted to try out various document servers and Wiki servers, including but not limited to Microsoft Search Server, WikiMedia Server and RedMine. At the end the client settled for WikiMedia Server and RedMine Server, hence the name WikiMine.